refactor(TableGroup at el): after separate packages, move entity, service, etc about TableGroup
